Fix Vercel preview prerender build
